encode/decode test pass
