+++ tests/data/storage/sample_0.json
@@ -1867,16 +1867,17 @@
           {"prim": "Pair",
            "args": [{"int": "0"},
             {"prim": "Pair",
-             "args": [{"bytes": "0000e26174b95dd7b766a1da439c36b3c9479dd9136f"},
+             "args": [{"string": "tz1gH29qAVaNfv7imhPthCwpUBcqmMdLWxPG"},
               {"prim": "Pair",
-               "args": [{"bytes": "0000e26174b95dd7b766a1da439c36b3c9479dd9136f"},
+               "args": [{"string": "tz1gH29qAVaNfv7imhPthCwpUBcqmMdLWxPG"},
                 {"prim": "Pair",
                  "args": [{"prim": "Pair",
                    "args": [{"prim": "True"},
                     {"prim": "Pair",
                      "args": [{"string": ""}, {"string": ""}]}]},
                   {"prim": "Pair",
-                   "args": [{"int": "0"}, {"int": "0"}]}]}]}]}]}]}]}]},
+                   "args": [{"string": "1970-01-01T00:00:00Z"},
+                    {"string": "1970-01-01T00:00:00Z"}]}]}]}]}]}]}]}]},
     {"prim": "Elt",
      "args": [{"int": "1"},
       {"prim": "Pair",
@@ -1886,9 +1887,9 @@
           {"prim": "Pair",
            "args": [{"int": "1000000"},
             {"prim": "Pair",
-             "args": [{"bytes": "00000347675c7f8278db9ef49cedc9721716a5d7e7c5"},
+             "args": [{"string": "tz1KwNFuzryhY7ZFyZVcdnHFEZCT9xgrRmWT"},
               {"prim": "Pair",
-               "args": [{"bytes": "0000e26174b95dd7b766a1da439c36b3c9479dd9136f"},
+               "args": [{"string": "tz1gH29qAVaNfv7imhPthCwpUBcqmMdLWxPG"},
                 {"prim": "Pair",
                  "args": [{"prim": "Pair",
                    "args": [{"prim": "False"},
@@ -1896,6 +1897,6 @@
                      "args": [{"string": "Moon Lander was damaged."},
                       {"string": "nil"}]}]},
                   {"prim": "Pair",
-                   "args": [{"int": "1555732800"},
-                    {"int": "1558324800"}]}]}]}]}]}]}]}]}],
+                   "args": [{"string": "2019-04-20T04:00:00Z"},
+                    {"string": "2019-05-20T04:00:00Z"}]}]}]}]}]}]}]}]}],
    {"int": "1"}]}}
